requirements.txt
asgiref==3.3.4
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.3
django-cors-headers==3.7.0
django-crontab==0.7.1
django-filter==2.4.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
Markdown==3.3.4
packaging==21.0
Pillow==8.2.0
psycopg2-binary==2.8.6
pyparsing==2.4.7
pyrfc==2.4.1
pytz==2021.1
requests==2.27.1
sqldiff==0.0.4
sqlparse==0.4.1
urllib3==1.26.9
uWSGI==2.0.19.1
